% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_plot_boxplots.R
\name{biomarker_plot_boxplots_server}
\alias{biomarker_plot_boxplots_server}
\title{Boxplots plot Server function}
\usage{
biomarker_plot_boxplots_server(id, calcVariableImportance, watermark = FALSE)
}
\arguments{
\item{id}{}
}
\description{
A shiny Module for plotting (server code).
}
