% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_plot_decisiontree.R
\name{biomarker_plot_decisiontree_server}
\alias{biomarker_plot_decisiontree_server}
\title{Decisiontree plot Server function}
\usage{
biomarker_plot_decisiontree_server(
  id,
  calcVariableImportance,
  watermark = FALSE
)
}
\arguments{
\item{id}{}
}
\description{
A shiny Module for plotting (server code).
}
