% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_plot_decisiontree.R
\name{biomarker_plot_decisiontree_ui}
\alias{biomarker_plot_decisiontree_ui}
\title{Decisiontree plot UI input function}
\usage{
biomarker_plot_decisiontree_ui(
  id,
  title,
  info.text,
  caption,
  label = "",
  height,
  width
)
}
\arguments{
\item{height}{}
}
\description{
A shiny Module for plotting (UI code).
}
