% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_plot_heatmap.R
\name{biomarker_plot_heatmap_server}
\alias{biomarker_plot_heatmap_server}
\title{Expression plot Server function}
\usage{
biomarker_plot_heatmap_server(
  id,
  calcVariableImportance,
  pgx,
  pdx_predicted,
  watermark = FALSE
)
}
\arguments{
\item{id}{}
}
\description{
A shiny Module for plotting (server code).
}
